requirements.txt
fastapi
uvicorn
fastapi-users[sqlalchemy]
pydantic
python-jose
python-dotenv
python-docx
jinja2
uvicorn
bcrypt==4.0.1
passlib[bcrypt]
arq
asyncpg
sqlalchemy[asyncio]
alembic
alembic_postgresql_enum
starlette
markdown
openai
langchain
langchain-openai
json_repair